Created listen_to_cells()
Previously talk_to_cell() both sent a message and got the reply, but that's not the way the real code will work.  The simulator will talk to individual cells and get messages from them, not necessarily as send/recv.  For example, the simulator may send chaos monkey messages to the cells, and the cells may send trace records to the simulator.  There is now a separate listen_to_cells() function that will listen to all the cells.