Got chaos monkey tests working
The test was working if both file handles were ready on the first call to select(), but not if only one was.  The only error message on the second call to select() was "EBAD", meaning I had a bad file handle.  It turns out that if you create a file handle from a raw file descriptor, an i32, that file handle calls close() when the owning variable goes out of scope.  That's certainly one definition of a bad file handle.

The solution was to create a struct to hold the file handle.  This version is the result of a long conversation with the borrow checker.  A struct Cell holds everything I need about a cell in order to talk to it over stdin/stdout.