Simplified a couple of calls
If a function/method takes an Option as an argument, you can write it in a way that the caller doesn't need to wrap the argument in Some.  That means you can call foo(None) or foo(x) vs. foo(Some(x)).  The called function/method will take care of it for you.

I also got rid of some (pun intended) compiler warnings.  (I actually got rid of all of them.)