Can run tests for multiprocess application
There is no standard framework for testing a multiprocess application in Rust, so I rolled my own.  Instead of starting the binaries for the worker processes, I run them under cargo test.  Then I collect the test reports from the workers to make sure they worked with an assert.

I would like to have separate reports for each worker separate from that of the main test, but for now I get a single report that only reports success if the main process and all workers succeed.  The main process does print the test reports from the workers, which can help.

One not so good thing is that the cargo test command builds the worker, which means it takes N builds to test with N workers.

NB: I'm committing this version even though all tests do not pass.