Getting close but not quite there
I can get either the messages from the cells or the cell test reports to the chaos monkey but not both.  This version uses pipes for the former and relies on capturing the cell stdout for the latter.  Unfortunately, I end up getting an empty string instead of the test report.  Other versions did get the test report.  You can run the past one or two versions to see that.

Sorry I didn't get a chance to finish.  Feel free to contact me if you have any questions.